requirements.txt
langchain
wikipedia
langchain-community
langchain-google-genai
python-dotenv
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv